tecnoParq/requirements.txt
# asgiref==3.2.10
# certifi==2020.6.20
# defusedxml==0.6.0
# diff-match-patch==20200713
# Django==3.1.6
# Pillow==8.1.2
# django-model-utils==4.1.1
# djangorestframework==3.11.2
# django-import-export==2.3.0
# jdcal==1.4.1
# odfpy==1.4.1
# openpyxl==3.0.4
# psycopg2==2.8.5
# sqlparse==0.3.1
# Unipath==1.1
# django-cors-headers==3.7.0
# django-filter==2.4.0
# boto3==1.17.53
# django-storages==1.11.1

asgiref==3.4.1
autopep8==1.6.0
boto3==1.19.8
botocore==1.22.8
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-model-utils==4.2.0
django-storages==1.12.3
djangorestframework==3.12.4
jmespath==0.10.0
# Markdown==3.3.4
Pillow==8.4.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
django-admin-rangefilter==0.9.0